chore(app): move file to satisfy serve message
