+++ svg/index.ts
